+++ jnetcap/impl/com/untangle/jnetcap/NetcapSession.java
@@ -1,22 +1,6 @@
-/*
- * $HeadURL$
- * Copyright (c) 2003-2007 Untangle, Inc. 
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License, version 2,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but
- * AS-IS and WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE, TITLE, or
- * NONINFRINGEMENT.  See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+/**
+ * $Id$
  */
-
-
 package com.untangle.jnetcap;
 
 import java.net.InetAddress;
@@ -54,7 +38,7 @@ public abstract class NetcapSession
 
 
     /* This is for children that override the SessionEndpoints class */
-    protected NetcapSession( int id, short protocol )
+    protected NetcapSession( long id, short protocol )
     {
         pointer = new CPointer( getSession( id, protocol ));
 
@@ -72,9 +56,9 @@ public abstract class NetcapSession
         return protocol;
     }
 
-    public int id() 
+    public long id() 
     {
-        return getIntValue( FLAG_ID, pointer.value());
+        return getLongValue( FLAG_ID, pointer.value());
     }
 
     public String toString( boolean ifClient )
@@ -218,7 +202,7 @@ public abstract class NetcapSession
     public Endpoints clientSide() { return clientSide; }
     public Endpoints serverSide() { return serverSide; }
 
-    private static native long getSession( int id, short protocol );
+    private static native long getSession( long id, short protocol );
     private static native void raze( long session );
     private static native String determineServerIntf( long session );
