+++ spyware/api/com/untangle/node/spyware/SpywareAccessEvent.java
@@ -78,12 +78,12 @@ public class SpywareAccessEvent extends SpywareEvent
      * @return the PipelineEndpoints.
      */
     @Column(name="session_id", nullable=false)
-    public Integer getSessionId()
+    public Long getSessionId()
     {
         return pipelineEndpoints.getSessionId();
     }
 
-    public void setSessionId( Integer sessionId )
+    public void setSessionId( Long sessionId )
     {
         this.pipelineEndpoints.setSessionId(sessionId);
     }
